Completed Functionality
All functionality is implemented and ready for full-scale testing

+++ marketplace.php
@@ -88,8 +88,6 @@ background-color:#4CAF50;
     </select>
       <input type="submit" value="Update"><br>
     </form>
-    
-    <h3><font color=#0099cc>Blueprints to purchase:</font></h3>
 <?php
 	session_start();
 	$inputPart;
@@ -135,7 +133,7 @@ background-color:#4CAF50;
 		$maxPrice = "";
 	}
 	else{
-		$maxPrice=addSlashes($_POST["price"]);
+		$maxPrice = round(addSlashes($_POST["price"]),2,PHP_ROUND_HALF_DOWN);
 		$checks++;
 	}
 	if($_POST["removeBuy"] != ""){
@@ -145,6 +143,22 @@ background-color:#4CAF50;
 	
 	$linkID = mysql_connect("localhost","jgavin","Furmanlax17");
 	mysql_select_db("jgavin", $linkID);
+	
+	$SQL = "SELECT Account_Balance,Available_Balance FROM Players WHERE Player_ID = ".$userID;
+	$allValues = mysql_query($SQL, $linkID);
+	if (!$allValues) {
+		echo "Could not successfully run query ($SQL) from DB: " . mysql_error();
+		exit;
+	}
+	echo "<TABLE BORDER=1 CELLPADDING=8>";
+	echo "<TR><TD><B>Current Balance</B></TD><TD><B>Available Balance</B></TD>";
+	$thisValue = mysql_fetch_assoc($allValues);
+	extract($thisValue);
+	echo "<TR>";
+	echo "<TD>\$".round($Account_Balance,2,PHP_ROUND_HALF_DOWN)."</TD>";
+	echo "<TD>\$".round($Available_Balance,2,PHP_ROUND_HALF_DOWN)."</TD>";
+	echo "</TR>";
+	echo "</TABLE>";
 
 
 	$SQL = "SELECT Team_Name FROM Team";
@@ -154,7 +168,7 @@ background-color:#4CAF50;
 		exit;
 	}
 	$totalrowsOverall = mysql_num_rows($allValues);	
-	$select= '<form action="marketplace.php" method="post">
+	$select= '<h3><font color=#0099cc>Blueprints to purchase:</font></h3><form action="marketplace.php" method="post">
 	<label for="teamBuy"><font color=#0099cc>Team To Purchase:</font></label>
 	<select name="teamBuy" id="teamBuy" title="teamBuy">';
 	for($i = 0;$i<$totalrowsOverall;$i++){
@@ -207,7 +221,7 @@ background-color:#4CAF50;
 				if($Price<=$maxPrice){
 					$sharesStillWanted = $numBuy - $sharesAquired;
 					//aquire the current balances of the buyer and seller
-					$SQL = "SELECT Account_Balance FROM Players WHERE Player_ID = '$Seller_ID'";
+					$SQL = "SELECT Account_Balance, Available_Balance FROM Players WHERE Player_ID = '$Seller_ID'";
 					$allValues = mysql_query($SQL, $linkID);
 					if (!$allValues) {
 						echo "Could not successfully run query ($SQL) from DB: " . mysql_error();
@@ -216,8 +230,9 @@ background-color:#4CAF50;
 					$thisValue = mysql_fetch_assoc($allValues);
 					extract($thisValue);
 					$sellerInitBal = $Account_Balance;
+					$sellerInitAvail = $Available_Balance;
 
-					$SQL = "SELECT Account_Balance FROM Players WHERE Player_ID = '$userID'";
+					$SQL = "SELECT Account_Balance, Available_Balance FROM Players WHERE Player_ID = '$userID'";
 					$allValues = mysql_query($SQL, $linkID);
 					if (!$allValues) {
 						echo "Could not successfully run query ($SQL) from DB: " . mysql_error();
@@ -226,10 +241,11 @@ background-color:#4CAF50;
 					$thisValue = mysql_fetch_assoc($allValues);
 					extract($thisValue);
 					$buyerInitBal = $Account_Balance;
+					$buyerInitAvail = $Available_Balance;
 					
 					//if the buyer can afford all wanted shares
 					$numSharesBuying = 0;
-					$buyerTestBal = $buyerInitBal;
+					$buyerTestBal = $buyerInitAvail;
 					for($i=0;$i<$sharesStillWanted;$i++){
 						$buyerTestBal = $buyerTestBal - $maxPrice;
 						if($buyerTestBal>0){
@@ -351,10 +367,13 @@ background-color:#4CAF50;
 					
 					//update both account balances and update transaction table
 					$sellerFinalBal = $sellerInitBal + ($Price * $numSharesBuying);
+					$sellerFinalAvail = $sellerInitAvail + ($Price * $numSharesBuying);
 					$buyerFinalBal = $buyerInitBal - ($Price * $numSharesBuying);
+					$buyerFinalAvail = $buyerInitAvail - ($Price * $numSharesBuying);
 
 					$SQL = "UPDATE Players 
-					SET Account_Balance = ".$sellerFinalBal."
+					SET Account_Balance = ".$sellerFinalBal.",
+					Available_Balance = ".$sellerFinalAvail."
 					WHERE Player_ID = ".$Seller_ID;
 					$allValues = mysql_query($SQL, $linkID);
 					if (!$allValues) {
@@ -362,7 +381,8 @@ background-color:#4CAF50;
 						exit;
 					}
 					$SQL = "UPDATE Players 
-					SET Account_Balance = ".$buyerFinalBal."
+					SET Account_Balance = ".$buyerFinalBal.",
+					Available_Balance = ".$buyerFinalAvail."
 					WHERE Player_ID = ".$userID;
 					$allValues = mysql_query($SQL, $linkID);
 					if (!$allValues) {
@@ -400,17 +420,36 @@ background-color:#4CAF50;
 				echo "Could not successfully run query ($SQL) from DB: " . mysql_error();
 				exit;
 			}
+			$totalCost = $maxPrice * $sharesStillWanted;
+			
+			$SQL = "SELECT Available_Balance FROM Players WHERE Player_ID = '$userID'";
+			$allValues = mysql_query($SQL, $linkID);
+			if (!$allValues) {
+				echo "Could not successfully run query ($SQL) from DB: " . mysql_error();
+				exit;
+			}
+			$thisValue = mysql_fetch_assoc($allValues);
+			extract($thisValue);
+			$availBalance = $Available_Balance - $totalCost;
+			
+			$SQL = "UPDATE Players SET Available_Balance = ".$availBalance." 
+			WHERE Player_ID = '$userID'";
+			$allValues = mysql_query($SQL, $linkID);
+			if (!$allValues) {
+				echo "Could not successfully run query ($SQL) from DB: " . mysql_error();
+				exit;
+			}
 		}
 	}
 	
 	
 	if($searchTerm==""){
-		$SQL = "SELECT te.Team_Name, bs.Amount_Selling, bs.Price, p.username AS Seller_ID 
+		$SQL = "SELECT te.Team_Name, bs.Amount_Selling, bs.Price 
 		FROM Players p, Team te, Blueprints_ForSale bs 
 		Where p.Player_ID = bs.Seller_ID and te.Team_ID = bs.Team_ID Order By ".$sort;
 	}
 	else{
-		$SQL = "SELECT te.Team_Name, bs.Amount_Selling, bs.Price, p.username AS Seller_ID 
+		$SQL = "SELECT te.Team_Name, bs.Amount_Selling, bs.Price
 		FROM Players p, Team te, Blueprints_ForSale bs Where p.Player_ID = bs.Seller_ID and 
 		te.Team_ID = bs.Team_ID and (Team_Name LIKE '%".$searchTerm."%' or 
 		Price LIKE '%".$searchTerm."%') ORDER BY ".$sort;
@@ -421,7 +460,7 @@ background-color:#4CAF50;
 		exit;
 	}
 	echo "<TABLE BORDER=1 CELLPADDING=8>";
-	echo "<TR><TD><B>Team_Name</B></TD><TD><B>Amount_Selling</B></TD><TD><B>Price</B></TD><TD><B>Seller_ID</B></TD>";
+	echo "<TR><TD><B>Team Name</B></TD><TD><B>Amount Selling</B></TD><TD><B>Price</B></TD>";
 		$totalrows = mysql_num_rows($allValues);
 		for ($i=1; $i <= $totalrows; $i++){
 			$thisValue = mysql_fetch_assoc($allValues);
@@ -429,8 +468,7 @@ background-color:#4CAF50;
 			echo "<TR>";
 			echo "<TD>$Team_Name</TD>";
 			echo "<TD>$Amount_Selling</TD>";
-			echo "<TD>\$$Price</TD>";
-			echo "<TD>$Seller_ID</TD>";
+			echo "<TD>\$".round($Price,2,PHP_ROUND_HALF_DOWN)."</TD>";
 			echo "</TR>";
 		}
 		echo "</TABLE>";
@@ -466,7 +504,7 @@ background-color:#4CAF50;
 	echo $select;
 
 	
-		$SQL = "SELECT te.Team_Name, bb.Amount_Buying, bb.Price, p.username AS Buyer_ID
+		$SQL = "SELECT te.Team_Name, bb.Amount_Buying, bb.Price
 		FROM Players p, Team te, BlueprintsToBuy bb
 		Where p.Player_ID = bb.buyerID and te.Team_ID = bb.Team_ID";
 	$allValues = mysql_query($SQL, $linkID);
@@ -475,7 +513,7 @@ background-color:#4CAF50;
 		exit;
 	}
   echo "<TABLE BORDER=1 CELLPADDING=8>";
-	echo "<TR><TD><B>Team_Name</B></TD><TD><B>Amount_Buying</B></TD><TD><B>Price</B></TD><TD><B>buyerID</B></TD>";
+	echo "<TR><TD><B>Team Name</B></TD><TD><B>Amount Buying</B></TD><TD><B>Price</B></TD>";
 		$totalrows = mysql_num_rows($allValues);
 		for ($i=1; $i <= $totalrows; $i++){
 			$thisValue = mysql_fetch_assoc($allValues);
@@ -483,8 +521,7 @@ background-color:#4CAF50;
 			echo "<TR>";
 			echo "<TD>$Team_Name</TD>";
 			echo "<TD>$Amount_Buying</TD>";
-			echo "<TD>\$$Price</TD>";
-			echo "<TD>$Buyer_ID</TD>";
+			echo "<TD>\$".round($Price,2,PHP_ROUND_HALF_DOWN)."</TD>";
 			echo "</TR>";
 		}
 		echo "</TABLE>";
@@ -494,6 +531,35 @@ background-color:#4CAF50;
 		function removeBuying($removeBuy,$userID){
 			$linkID = mysql_connect("localhost","jgavin","Furmanlax17");
 			mysql_select_db("jgavin", $linkID);
+			$SQL = "SELECT Available_Balance FROM Players WHERE Player_ID = '$userID'";
+			$allValues = mysql_query($SQL, $linkID);
+			if (!$allValues) {
+				echo "Could not successfully run query ($SQL) from DB: " . mysql_error();
+				exit;
+			}
+			$thisValue = mysql_fetch_assoc($allValues);
+			extract($thisValue);
+			$availBalance = $Available_Balance;
+			
+			$SQL = "SELECT Price, Amount_Buying FROM BlueprintsToBuy 
+			WHERE buyerID = '$userID' AND Team_ID = (SELECT Team_ID FROM Team WHERE Team_Name = '$removeBuy')";
+			$allValues = mysql_query($SQL, $linkID);
+			if (!$allValues) {
+				echo "Could not successfully run query ($SQL) from DB: " . mysql_error();
+				exit;
+			}
+			$thisValue = mysql_fetch_assoc($allValues);
+			extract($thisValue);
+			$totalCost = $Price * $Amount_Buying;
+			$availBalance = $availBalance + $totalCost;
+			
+			$SQL = "UPDATE Players SET Available_Balance = ".$availBalance." 
+			WHERE Player_ID = '$userID'";
+			$allValues = mysql_query($SQL, $linkID);
+			if (!$allValues) {
+				echo "Could not successfully run query ($SQL) from DB: " . mysql_error();
+				exit;
+			}
 			
 			$SQL = "DELETE FROM BlueprintsToBuy WHERE buyerID = '$userID' 
 			AND Team_ID = (SELECT Team_ID FROM Team WHERE Team_Name = '$removeBuy')";
